feat: no more classes
