--- src/WebClient/app/types/index.ts
@@ -1,1 +1,0 @@
-export * from '@microsoftgraveyard/types/corpse';
